refactor(guest-agent): share system log test guard
