+++ crates/guest-agent/tests/claude_result_failure_logging.rs
@@ -6,25 +6,10 @@
 mod common;
 
 use agent_diagnostics::FailureDetailSource;
+use common::SystemLogOverrideGuard;
 use guest_agent::http::HttpClient;
 use guest_agent::masker::SecretMasker;
-use std::path::Path;
 use std::time::Duration;
-
-struct SystemLogOverrideGuard;
-
-impl SystemLogOverrideGuard {
-    fn set(path: &Path) -> Self {
-        guest_common::log::set_system_log_file(path);
-        Self
-    }
-}
-
-impl Drop for SystemLogOverrideGuard {
-    fn drop(&mut self) {
-        guest_common::log::clear_system_log_file();
-    }
-}
 
 #[tokio::test]
 async fn claude_error_result_is_written_to_system_log() -> Result<(), Box<dyn std::error::Error>> {
