+++ crates/guest-agent/tests/integration/mod.rs
@@ -3,6 +3,9 @@
 // serialises correctly (each runtime owns its own OS thread).
 #![allow(clippy::await_holding_lock)]
 
+#[path = "../common/mod.rs"]
+mod common;
+
 #[macro_use]
 mod support;
